requirements.txt
Django
python-dotenv
requests
stripe
celery
redis
gunicorn
